style: 스캔 상단 탭 스타일 수정
