+++ apps/web/src/app/wrong/create/scans/page.tsx
@@ -1,15 +1,12 @@
 ﻿"use client";
 
 import { useEffect, useMemo, useState } from "react";
-import Image from "next/image";
-import Link from "next/link";
 import clsx from "clsx";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useCompleteWrongCreateScansMutation } from "@/shared/apis/problem-create/hooks/use-complete-wrong-create-scans-mutation";
 import { Button } from "@/shared/components/button/button/button";
-import Chip from "@/shared/components/chip/chip";
 import Divider from "@/shared/components/divider/divider";
-import Icon from "@/shared/components/icon/icon";
+import ScanCard from "@/shared/components/scan-card/scan-card";
 import CompleteModal from "@/shared/components/modal/complete-modal/complete-modal";
 import { ROUTES } from "@/shared/constants/routes";
 import {
@@ -138,47 +135,15 @@ const WrongCreateScansPage = () => {
         ) : (
           <div className={s.grid}>
             {visibleItems.map((item) => (
-              <div key={item.scanId} className={s.card}>
-                <Link href={buildProblemHref(item)} className={s.cardLink}>
-                  <div className={s.cardFrame}>
-                    {item.imageUrl ? (
-                      <Image
-                        src={item.imageUrl}
-                        alt={item.title}
-                        fill
-                        unoptimized
-                        className={s.cardImage}
-                      />
-                    ) : null}
-                    <div className={s.cardOverlay} />
-                    <div className={s.cardBody}>
-                      <div className={s.cardContent}>
-                        <div className={s.chipWrap}>
-                          <Chip
-                            as="span"
-                            label={item.unitName}
-                            size="xs"
-                            shape="square"
-                            tone="surface"
-                            className={s.unitChip}
-                          />
-                        </div>
-                        <p className={s.cardTitle}>{item.title}</p>
-                      </div>
-                    </div>
-                  </div>
-
-                  <span className={s.subjectChip}>{item.subjectName}</span>
-                </Link>
-                <button
-                  type="button"
-                  className={s.deleteButton}
-                  aria-label={`${item.title} 삭제`}
-                  onClick={() => setDeleteTarget(item)}
-                >
-                  <Icon name="trash-chip" size={1.8} />
-                </button>
-              </div>
+              <ScanCard
+                key={item.scanId}
+                title={item.title}
+                subjectName={item.subjectName}
+                unitNames={[item.unitName]}
+                imageSrc={item.imageUrl}
+                href={buildProblemHref(item)}
+                onDelete={() => setDeleteTarget(item)}
+              />
             ))}
           </div>
         )}
